pyproject.toml
[tool.poetry]
name = "elo-indexserver-client"
version = "0.1.9"
description = "A client library for accessing indexserver REST API of the ECM system ELO Digital Office."

authors = [
    "Manuel Eiweck <manuel.eiweck@treskon.at>"
]

maintainers = [
    "Treskon GmbH <internal-dev@treskon.at>",
    "Manuel Eiweck <manuel.eiweck@treskon.at>"
]
documentation = "https://install.portrait.app/elo-indexserver-client/html/"

readme = "README.md"
packages = [
    {include = "eloclient"},
    {include = "eloservice"}
]
include = ["CHANGELOG.md", "eloclient/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.26.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[tool.poetry.dev-dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
